performance improvements + renamed score columns in results

+++ RetrievalAlgorithm/src/unimodal_calculations.py
@@ -57,7 +57,7 @@ def calculate_unimodal_similarity(
                 feature_1_batch = feature_1_batch.to(device)
                 feature_2_batch = feature_2_batch.to(device)
 
-                with torch.amp.autocast(device_type=device.type, dtype=torch.float16):
+                with torch.amp.autocast(device_type=device.type):
                     feature_1_batch = normalization_module(feature_1_batch)
                     feature_2_batch = normalization_module(feature_2_batch)
                     sim_scores_batch = calculation_module(feature_1_batch, feature_2_batch).view(-1)
